# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/brokermr810/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/market/config:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Public config for frontend (local mode).
      tags:
      - Market
      operationId: getApiMarketConfig
      description: 'Public config for frontend (local mode).

        Mirrors the old PHP `/addons/quantdinger/index/getConfig` shape.'
      x-visibility: public
  /api/market/types:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Return supported market types for the add-watchlist modal.
      description: 'Visibility is controlled by the centralised :mod:`app.utils.market_visibility`

        rules, primarily the ``ENABLED_MARKETS`` whitelist, with

        ``SHOW_CN_STOCK`` / ``SHOW_HK_STOCK`` kept for back-compat. The radar

        (``/api/global-market/opportunities``) and the Agent API

        (``/api/agent/v1/markets``) read from the same helper so all three

        user-facing market lists stay in lock-step.'
      tags:
      - Market
      operationId: getApiMarketTypes
      x-visibility: public
  /api/market/menuFooterConfig:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Compatibility stub for old PHP `getMenuFooterConfig`.
      tags:
      - Market
      operationId: getApiMarketMenuFooterConfig
      description: 'Compatibility stub for old PHP `getMenuFooterConfig`.

        Frontend can also hardcode this locally; this endpoint remains for completeness.'
      x-visibility: public
  /api/market/symbols/search:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Lightweight symbol search.
      tags:
      - Market
      operationId: getApiMarketSymbolsSearch
      description: 'Lightweight symbol search.

        DB seed first; for Crypto, falls back to exchange market list when DB yields
        few results.'
      x-visibility: public
  /api/market/symbols/hot:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Return a small curated hot list per market (local-only).
      tags:
      - Market
      operationId: getApiMarketSymbolsHot
      x-visibility: public
  /api/market/watchlist/get:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get watchlist for the current user.
      tags:
      - Market
      operationId: getApiMarketWatchlistGet
      x-visibility: public
  /api/market/watchlist/add:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Add a symbol to watchlist for the current user.
      tags:
      - Market
      operationId: postApiMarketWatchlistAdd
      x-visibility: public
  /api/market/watchlist/remove:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Remove a symbol from watchlist for the current user.
      tags:
      - Market
      operationId: postApiMarketWatchlistRemove
      x-visibility: public
  /api/market/watchlist/prices:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get realtime prices for the current user's watchlist.
      tags:
      - Market
      operationId: getApiMarketWatchlistPrices
      x-visibility: public
  /api/market/price:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get realtime price for a single symbol.
      description: 'Query params:

        market: Market type

        symbol: Symbol or ticker'
      tags:
      - Market
      operationId: getApiMarketPrice
      x-visibility: public
...
openapi: 3.0.3